INT-191 introduce buildInferenceJobs
