+++ src/components/jobManager.ts
@@ -29,6 +29,7 @@ import {destructIntuitaFileSystemUri} from "../destructIntuitaFileSystemUri";
 import { extractKindsFromTs2345ErrorMessage } from "../features/repairCode/extractKindsFromTs2345ErrorMessage";
 import { buildReplacement } from "../features/repairCode/buildReplacement";
 import { InferenceJob } from "./inferenceService";
+import { buildInferenceJobs } from "../features/repairCode/buildInferenceJobs";
 
 type Job = MoveTopLevelNodeJob | RepairCodeJob;
 
@@ -343,44 +344,7 @@ export class JobManager {
     public onRuleBasedCoreRepairDiagnosticsChanged(
         message: Message & { kind: MessageKind.ruleBasedCoreRepairDiagnosticsChanged },
     ) {
-        const separator = getSeparator(message.text);
-        const lines = calculateLines(message.text, separator);
-        const lengths = calculateLengths(lines);
-
-        const inferenceJobs: ReadonlyArray<InferenceJob> = message.diagnostics
-            .map((diagnostic) => {
-                const kinds = extractKindsFromTs2345ErrorMessage(diagnostic.message);
-
-                if(!kinds) {
-                    return null;
-                }
-
-                const range: IntuitaRange = [
-                    diagnostic.range.start.line,
-                    diagnostic.range.start.character,
-                    diagnostic.range.end.line,
-                    diagnostic.range.end.character,
-                ];
-
-                const intuitaSimpleRange = buildIntuitaSimpleRange(
-                    separator,
-                    lengths,
-                    range,
-                );
-
-                const rangeText = message.text.slice(
-                    intuitaSimpleRange.start,
-                    intuitaSimpleRange.end,
-                );
-
-                const replacement = buildReplacement(rangeText, kinds.expected);
-
-                return {
-                    range,
-                    replacement,
-                };
-            })
-            .filter(isNeitherNullNorUndefined);
+        const inferenceJobs = buildInferenceJobs(message);
 
         const fileUri = buildFileUri(message.uri);
